# examples/todoApp/open_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "An API for managing tasks, colors, statuses, and changelogs."
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local server"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "x-celestial-grouping": "tasks",
        "x-celestial-index-endpoint": true,
        "tags": ["Tasks"],
        "description": "Returns a list of all tasks",
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": ["Tasks"],
        "x-celestial-grouping": "tasks",
        "description": "Adds a new task to the database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task": {
        "delete": {
            "summary": "Delete a task",
            "tags": ["Tasks"],
            "x-celestial-grouping": "tasks",
            "x-celestial-updateByKey": "requestBody.task_id",
            "description": "Deletes a task from the database",
            "requestBody": {
                "required": true,
                "content": {
                "application/json": {
                    "schema": {
                    "$ref": "#/components/schemas/DeleteTask"
                    }
                }
                }
            },
            "responses": {
                "200": {
                "description": "Task deleted successfully",
                "content": {
                    "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/DeleteResponse"
                    }
                    }
                }
                }
            }
        }
    },
    "/task/color": {
      "put": {
        "summary": "Update task color",
        "tags": ["Tasks"],
        "x-celestial-grouping": "tasks",
        "x-celestial-updateByKey": "requestBody.task_id",
        "description": "Updates the color of an existing task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskColor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task color updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/task/status": {
      "put": {
        "summary": "Update task status",
        "tags": ["Tasks"],
        "x-celestial-grouping": "tasks",
        "x-celestial-updateByKey": "requestBody.task_id",
        "description": "Updates the status of an existing task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/colors": {
      "get": {
        "summary": "Get all colors",
        "tags": ["Colors"],
        "x-celestial-grouping": "colors",
        "x-celestial-index-endpoint": true,
        "description": "Returns a list of all colors",
        "responses": {
          "200": {
            "description": "A list of colors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Color"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "summary": "Get all statuses",
        "tags": ["Status"],
        "x-celestial-grouping": "status",
        "x-celestial-index-endpoint": true,
        "description": "Returns a list of all statuses",
        "responses": {
          "200": {
            "description": "A list of statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Status"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "text": {
            "type": "string",
            "example": "Example task"
          },
          "color": {
            "type": "integer",
            "example": 2
          },
          "status": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "NewTask": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Example task"
          }
        },
        "required": ["text"]
      },
      "UpdateTaskColor": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "example": 1
          },
          "color": {
            "type": "integer",
            "example": 2
          }
        },
        "required": ["task_id", "color"]
      },
      "UpdateTaskStatus": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "integer",
            "example": 1
          }
        },
        "required": ["task_id", "status"]
      },
      "DeleteTask": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "example": 1
          }
        },
        "required": ["task_id"]
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Task added/updated successfully"
          },
          "inserted": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Task deleted successfully"
          },
          "inserted": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              }
            }
          }
        }
      },
      "Color": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "color": {
            "type": "string",
            "example": "Red"
          }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "example": "Completed"
          }
        }
      }
    }
  }
}